LayoutTests/imported/w3c/web-platform-tests/css/css-ui/text-overflow-023-expected.txt

    

FAIL Checks hit testing on the ellipsis assert_equals: the element targeted by a hit on the ellipsis is the elided inline. expected "target" but got "parent"
